+++ catboost/node-package/.npmignore
@@ -1,3 +1,5 @@
+bindings/**/*
 build/**/*
 test_data/**/*
 lib/tests/**/*
+tsconfig.json
