+++ catboost/node-package/src/model.cpp
@@ -5,6 +5,7 @@
 
 namespace {
 
+// Collect pointers to matrix rows into a vector.
 template <typename T, typename V = const T*, typename C = const TVector<T>>
 TVector<V> CollectMatrixRowPointers(C& matrix, uint32_t rowLength) {
     TVector<V> pointers;
@@ -15,31 +16,20 @@ TVector<V> CollectMatrixRowPointers(C& matrix, uint32_t rowLength) {
     return pointers;
 }
 
-template <typename T>
-Napi::Array ConvertToArray(Napi::Env env, const TVector<T>& values) {
-    Napi::Array result = Napi::Array::New(env);
-    uint32_t index = 0;
-    for (const auto value: values) {
-        result[index++] = Napi::Number::New(env, value);
-    }
-
-    return result;
 }
-
-} 
 
 namespace NNodeCatBoost {
 
 TModel::TModel(const Napi::CallbackInfo& info): Napi::ObjectWrap<TModel>(info) {
     Napi::Env env = info.Env();
 
-    this->Handle = ModelCalcerCreate(); 
+    this->Handle = ModelCalcerCreate();
     NHelper::CheckNotNullHandle(env, this->Handle);
 }
 
 TModel::~TModel() {
     if (this->Handle != nullptr) {
-       ModelCalcerDelete(this->Handle); 
+       ModelCalcerDelete(this->Handle);
     }
 }
 
@@ -62,7 +52,7 @@ void TModel::LoadFullFromFile(const Napi::CallbackInfo& info) {
     NHelper::Check(env, info[0].IsString(), "File name string is required");
 
     NHelper::CheckNotNullHandle(env, this->Handle);
-    const bool status = LoadFullModelFromFile(this->Handle, 
+    const bool status = LoadFullModelFromFile(this->Handle,
                                               info[0].As<Napi::String>().Utf8Value().c_str());
     NHelper::CheckStatus(env, status);
 }
@@ -71,7 +61,7 @@ Napi::Value TModel::CalcPrediction(const Napi::CallbackInfo& info) {
     Napi::Env env = info.Env();
 
     NHelper::Check(env, info.Length() >= 2, "Wrong number of arguments - expected 2");
-    NHelper::Check(env, NHelper::IsMatrix(info[0], NHelper::NAT_NUMBER),  
+    NHelper::Check(env, NHelper::IsMatrix(info[0], NHelper::NAT_NUMBER),
         "Expected the first argument to be a matrix of floats");
 
     const Napi::Array floatFeatures = info[0].As<Napi::Array>();
@@ -81,7 +71,7 @@ Napi::Value TModel::CalcPrediction(const Napi::CallbackInfo& info) {
     }
 
     const uint32_t floatFeaturesSize = floatFeatures[0u].As<Napi::Array>().Length();
-    
+
     TVector<float> floatFeatureValues;
     floatFeatureValues.reserve(floatFeaturesSize * docsCount);
 
@@ -92,9 +82,9 @@ Napi::Value TModel::CalcPrediction(const Napi::CallbackInfo& info) {
         }
     }
 
-    NHelper::Check(env, 
+    NHelper::Check(env,
         NHelper::IsMatrix(info[1], NHelper::NAT_NUMBER) || NHelper::IsMatrix(info[1], NHelper::NAT_STRING),
-        "Expected second argument to be a matrix of strings or numbers");    
+        "Expected second argument to be a matrix of strings or numbers");
     const Napi::Array catFeatures = info[1].As<Napi::Array>();
 
     NHelper::Check(env, catFeatures.Length() == docsCount,
@@ -134,8 +124,8 @@ Napi::Value TModel::GetModelDimensionsCount(const Napi::CallbackInfo& info) {
     return Napi::Number::New(env, count);
 }
 
-Napi::Array TModel::CalcPredictionHash(Napi::Env env, 
-                                   const TVector<float>& floatFeatures, 
+Napi::Array TModel::CalcPredictionHash(Napi::Env env,
+                                   const TVector<float>& floatFeatures,
                                    const Napi::Array& catFeatures) {
     const uint32_t docsCount = catFeatures.Length();
     const uint32_t catFeaturesSize = catFeatures[0u].As<Napi::Array>().Length();
@@ -156,17 +146,17 @@ Napi::Array TModel::CalcPredictionHash(Napi::Env env,
 
     TVector<const float*> floatPtrs = CollectMatrixRowPointers<float>(floatFeatures, floatFeaturesSize);
     TVector<const int*> catPtrs = CollectMatrixRowPointers<int>(catHashValues, catFeaturesSize);
-    NHelper::CheckStatus(env, 
-        CalcModelPredictionWithHashedCatFeatures(this->Handle, docsCount, 
+    NHelper::CheckStatus(env,
+        CalcModelPredictionWithHashedCatFeatures(this->Handle, docsCount,
                         floatPtrs.data(), floatFeaturesSize,
                         catPtrs.data(), catFeaturesSize,
                         resultValues.data(), docsCount));
 
-    return ConvertToArray(env, resultValues);
+    return NHelper::ConvertToArray(env, resultValues);
 }
 
-Napi::Array TModel::CalcPredictionString(Napi::Env env, 
-                                   const TVector<float>& floatFeatures, 
+Napi::Array TModel::CalcPredictionString(Napi::Env env,
+                                   const TVector<float>& floatFeatures,
                                    const Napi::Array& catFeatures) {
     const uint32_t docsCount = catFeatures.Length();
     const uint32_t catFeaturesSize = catFeatures[0u].As<Napi::Array>().Length();
@@ -190,13 +180,13 @@ Napi::Array TModel::CalcPredictionString(Napi::Env env,
 
     TVector<const float*> floatPtrs = CollectMatrixRowPointers<float>(floatFeatures, floatFeaturesSize);
     TVector<const char**> catPtrs = CollectMatrixRowPointers<const char*, const char**>(catStringValues, catFeaturesSize);
-    NHelper::CheckStatus(env, 
-        CalcModelPrediction(this->Handle, docsCount, 
+    NHelper::CheckStatus(env,
+        CalcModelPrediction(this->Handle, docsCount,
                         floatPtrs.data(), floatFeaturesSize,
                         catPtrs.data(), catFeaturesSize,
                         resultValues.data(), docsCount));
-    
-    return ConvertToArray(env, resultValues);
+
+    return NHelper::ConvertToArray(env, resultValues);
 }
 
 }
